chore: remove synchronize lock
