--- composeApp/src/commonMain/kotlin/org/ooni/probe/CreateDataStore.kt
@@ -1,34 +1,0 @@
-package org.ooni.probe
-
-import androidx.datastore.core.DataMigration
-import androidx.datastore.core.DataStore
-import androidx.datastore.preferences.core.PreferenceDataStoreFactory
-import androidx.datastore.preferences.core.Preferences
-import kotlinx.atomicfu.locks.SynchronizedObject
-import kotlinx.atomicfu.locks.synchronized
-import okio.Path.Companion.toPath
-
-private lateinit var dataStore: DataStore<Preferences>
-
-private val lock = SynchronizedObject()
-
-/**
- * Gets the singleton DataStore instance, creating it if necessary.
- */
-fun getDataStore(
-    producePath: () -> String,
-    migrations: List<DataMigration<Preferences>> = listOf(),
-): DataStore<Preferences> =
-    synchronized(lock) {
-        if (::dataStore.isInitialized) {
-            dataStore
-        } else {
-            PreferenceDataStoreFactory.createWithPath(
-                produceFile = { producePath().toPath() },
-                migrations = migrations,
-            )
-                .also { dataStore = it }
-        }
-    }
-
-internal const val DATA_STORE_FILE_NAME = "probe.preferences_pb"
