+++ composeApp/src/iosTest/kotlin/org/ooni/testing/CreatePreferenceDataStore.kt
@@ -2,8 +2,7 @@ package org.ooni.testing
 
 import androidx.datastore.core.DataStore
 import androidx.datastore.preferences.core.Preferences
-import org.ooni.probe.DATA_STORE_FILE_NAME
-import org.ooni.probe.getDataStore
+import org.ooni.probe.di.Dependencies
 import platform.Foundation.NSDocumentDirectory
 import platform.Foundation.NSFileManager
 import platform.Foundation.NSURL
@@ -11,7 +10,7 @@ import platform.Foundation.NSUserDomainMask
 
 @OptIn(kotlinx.cinterop.ExperimentalForeignApi::class)
 internal actual fun createPreferenceDataStore(): DataStore<Preferences> {
-    return getDataStore(
+    return Dependencies.getDataStore(
         producePath = {
             val documentDirectory: NSURL? =
                 NSFileManager.defaultManager.URLForDirectory(
@@ -21,7 +20,7 @@ internal actual fun createPreferenceDataStore(): DataStore<Preferences> {
                     create = false,
                     error = null,
                 )
-            requireNotNull(documentDirectory).path + "/$DATA_STORE_FILE_NAME"
+            requireNotNull(documentDirectory).path + "/${Dependencies.Companion.DATA_STORE_FILE_NAME}"
         },
     )
 }
