+++ gradle/libs.versions.toml
@@ -66,7 +66,6 @@ ui = [
 tooling = [
     "kermit",
     "sqldelight-coroutines",
-    "kotlinx-atomicfu",
     "androidx-datastore-core-okio",
     "androidx-datastore-preferences-core",
 ]
